backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
python-decouple==3.8
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
drf-spectacular==0.28.0
sentry-sdk[django]==2.19.2
python-json-logger==3.2.1

# Google AI providers
google-genai==0.3.0
google-cloud-aiplatform==1.74.0
google-auth==2.36.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

# Dev tools
ipython==8.30.0
